% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{autosomes}
\alias{autosomes}
\title{Return subset of autosomes.}
\usage{
autosomes(this_table)
}
\arguments{
\item{this_table}{A character.}
}
\value{
data.frame
}
\description{
Return subset of autosomes.
}
\seealso{
Other utility functions: \code{\link{get_grch_gene_table}},
  \code{\link{get_meth_table}},
  \code{\link{get_ucsc_refgene_table}},
  \code{\link{longest_refgene_transcripts}},
  \code{\link{normal_chroms}}
}

