% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{genomic_complement}
\alias{genomic_complement}
\title{Get the rows from table_1 with coords that don't intersect the coords in table_2}
\usage{
genomic_complement(table_1, table_2)
}
\arguments{
\item{table_1}{A data.table in BED format.}

\item{table_2}{A data.table in BED format.}
}
\value{
A data.table.
}
\description{
Get the rows from table_1 with coords that don't intersect the coords in table_2
}
\seealso{
Other workorse functions: \code{\link{get_bin_ranges_bed}},
  \code{\link{get_bin_ranges_refgene}},
  \code{\link{get_binned_perc_meth_list}},
  \code{\link{get_binned_perc_meth}},
  \code{\link{get_genome_range_overlaps}},
  \code{\link{get_intersect_single_bases}},
  \code{\link{parallel_genomic_complement}}
}

