% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_aggregate_bed_enrichment_over_ranges}
\alias{get_aggregate_bed_enrichment_over_ranges}
\title{Get overlap enrichment in ranges for generic BED-formatted file (e.g. peak ranges)}
\usage{
get_aggregate_bed_enrichment_over_ranges(ranges, bed_path)
}
\arguments{
\item{ranges}{A data.frame or data table.}

\item{bed_path}{A character.}
}
\value{
data.table
}
\description{
Get overlap enrichment in ranges for generic BED-formatted file (e.g. peak ranges)
}
\seealso{
Other workorse functions: \code{\link{add_chip_enrich_to_binned_perc_meth}};
  \code{\link{genomic_complement}};
  \code{\link{get_bin_ranges_bed}};
  \code{\link{get_bin_ranges_refgene}};
  \code{\link{get_binned_perc_meth_list}};
  \code{\link{get_binned_perc_meth}};
  \code{\link{get_genome_range_overlaps}};
  \code{\link{get_intersect_single_bases}};
  \code{\link{parallel_genomic_complement}}
}

