% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_bin_ranges_bed}
\alias{get_bin_ranges_bed}
\title{Get ranges of bins from BED element centers.}
\usage{
get_bin_ranges_bed(bed_table_path, bin_width, range)
}
\arguments{
\item{bed_table_path}{A character.}

\item{bin_width}{A numeric.}

\item{range}{A numeric.}
}
\value{
data.table
}
\description{
Get ranges of bins from BED element centers.
}
\seealso{
Other workorse functions: \code{\link{add_chip_enrich_to_binned_perc_meth}},
  \code{\link{genomic_complement}},
  \code{\link{get_aggregate_bed_enrichment_over_ranges}},
  \code{\link{get_bin_ranges_refgene}},
  \code{\link{get_binned_perc_meth_list}},
  \code{\link{get_binned_perc_meth}},
  \code{\link{get_genome_range_overlaps}},
  \code{\link{get_intersect_single_bases}},
  \code{\link{parallel_genomic_complement}}
}

