% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_binned_perc_meth}
\alias{get_binned_perc_meth}
\title{Get percent methylation table by bin.}
\usage{
get_binned_perc_meth(ranges, meth_table)
}
\arguments{
\item{ranges}{A data.frame.}

\item{meth_table}{A data.frame.}
}
\value{
data.table
}
\description{
Get percent methylation table by bin.
}
\seealso{
Other workorse functions: \code{\link{get_bin_ranges_bed}};
  \code{\link{get_bin_ranges_refgene}}
}

