% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_binned_perc_meth_list}
\alias{get_binned_perc_meth_list}
\title{Get percent methylation table by bin for list of meth tables.}
\usage{
get_binned_perc_meth_list(bin_ranges, meth_table_list, cores = 1)
}
\arguments{
\item{bin_ranges}{A data.frame.}

\item{meth_table_list}{A list of data.tables.}

\item{cores}{An integer numeric.}
}
\value{
list
}
\description{
Get percent methylation table by bin for list of meth tables.
}
\seealso{
Other workorse functions: \code{\link{genomic_complement}},
  \code{\link{get_bin_ranges_bed}},
  \code{\link{get_bin_ranges_refgene}},
  \code{\link{get_binned_perc_meth}},
  \code{\link{get_genome_range_overlaps}},
  \code{\link{get_intersect_single_bases}},
  \code{\link{parallel_genomic_complement}}
}

