% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_gene_quantiles}
\alias{get_gene_quantiles}
\title{Get gene quantiles for single gene.}
\usage{
get_gene_quantiles(trancript_line, quantiles = 100)
}
\arguments{
\item{trancript_line}{A data.frame or data table .}

\item{quantiles}{A numeric.}
}
\value{
data.table
}
\description{
Get gene quantiles for single gene.
}
\seealso{
Other plotting functions: \code{\link{plot_generic_aggregate_enrichment}},
  \code{\link{plot_percent_meth_with_depth}},
  \code{\link{plot_percent_meth}}
}

