% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_genome_range_overlaps}
\alias{get_genome_range_overlaps}
\title{Get range overlaps}
\usage{
get_genome_range_overlaps(bin_ranges, bed_path, cores = 1)
}
\arguments{
\item{bin_ranges}{A data.frame.}

\item{bed_path}{A character.}

\item{cores}{An integer.}
}
\value{
data.table
}
\description{
Get range overlaps
}
\seealso{
Other workorse functions: \code{\link{get_bin_ranges_bed}};
  \code{\link{get_bin_ranges_refgene}};
  \code{\link{get_binned_perc_meth}};
  \code{\link{get_intersect_single_bases}}
}

