% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_grch_gene_table}
\alias{get_grch_gene_table}
\title{Import and format grch37 gene model file.}
\usage{
get_grch_gene_table(grch37_table_path)
}
\arguments{
\item{grch37_table_path}{A character.}
}
\value{
data.frame
}
\description{
Import and format grch37 gene model file.
}
\seealso{
Other utility functions: \code{\link{autosomes}};
  \code{\link{get_meth_table}};
  \code{\link{get_ucsc_refgene_table}};
  \code{\link{longest_refgene_transcripts}};
  \code{\link{normal_chroms}}
}

