% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_meth_table}
\alias{get_meth_table}
\title{Import and format methyl sequencing file.}
\usage{
get_meth_table(meth_path)
}
\arguments{
\item{meth_path}{A character.}
}
\value{
data.table
}
\description{
Import and format methyl sequencing file.
}
\seealso{
Other utility functions: \code{\link{autosomes}},
  \code{\link{get_grch_gene_table}},
  \code{\link{get_ucsc_refgene_table}},
  \code{\link{longest_refgene_transcripts}},
  \code{\link{normal_chroms}}
}

