% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_ucsc_refgene_table}
\alias{get_ucsc_refgene_table}
\title{Import and format UCSC refgene gene model file.}
\usage{
get_ucsc_refgene_table(refgene_path)
}
\arguments{
\item{refgene_path}{A character.}
}
\value{
data.frame
}
\description{
Import and format UCSC refgene gene model file.
}
\seealso{
Other utility functions: \code{\link{autosomes}};
  \code{\link{get_grch_gene_table}};
  \code{\link{get_meth_table}};
  \code{\link{longest_refgene_transcripts}};
  \code{\link{normal_chroms}}
}

