% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methplot.r
\name{longest_refgene_transcripts}
\alias{longest_refgene_transcripts}
\title{Get longest possible gene model from refgene.}
\usage{
longest_refgene_transcripts(ref_table)
}
\arguments{
\item{ref_table}{A data.frame or data.table.}
}
\value{
data.table or data.frame
}
\description{
Get longest possible gene model from refgene.
}
\seealso{
Other utility functions: \code{\link{autosomes}};
  \code{\link{get_grch_gene_table}};
  \code{\link{get_meth_table}};
  \code{\link{get_ucsc_refgene_table}};
  \code{\link{normal_chroms}}
}

