\name{mydata}
\docType{data}
\alias{mydata}
\title{FOXP3 Methylation profiles of two samples}
\description{
      This data set gives the read number detected in each methylation pattern in FOXP3 T-cell specific demthylation region using next generation bisulphite sequencing. Two samples names "P1_A1" and "P1_A2" are given. 
}
\usage{mydata}
\format{A matrix with 1024 rows and 2 columns where rows indicate how many reads detected in each methylation pattern and columns stand for two samples.}
\keyword{datasets}


