% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methplot.r
\name{normal_chroms}
\alias{normal_chroms}
\title{Return subset of autosomes plus X and Y.}
\usage{
normal_chroms(this_table)
}
\arguments{
\item{this_table}{A character.}
}
\value{
data.frame
}
\description{
Return subset of autosomes plus X and Y.
}
\seealso{
Other utility functions: \code{\link{autosomes}};
  \code{\link{get_grch_gene_table}};
  \code{\link{get_meth_table}};
  \code{\link{get_ucsc_refgene_table}};
  \code{\link{longest_refgene_transcripts}}
}

