% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methplot.r
\name{parallel_genomic_intersect}
\alias{parallel_genomic_intersect}
\title{Quickly get the overlaps between two bed-like tables via data splitting and parallel processing}
\usage{
parallel_genomic_intersect(table_1, table_2, cores = 1)
}
\arguments{
\item{table_1}{A data.table.}

\item{table_2}{A data.table.}

\item{cores}{An integer numeric.}
}
\value{
data.table
}
\description{
Quickly get the overlaps between two bed-like tables via data splitting and parallel processing
}

