% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{plot_generic_aggregate_enrichment}
\alias{plot_generic_aggregate_enrichment}
\title{Plot binned aggregate enrichments arround genomic-range-defined (BED-defined) centers.}
\usage{
plot_generic_aggregate_enrichment(ranges, bed_path)
}
\arguments{
\item{ranges}{A data.frame or data table.}

\item{bed_path}{A character.}
}
\value{
ggplot
}
\description{
Plot binned aggregate enrichments arround genomic-range-defined (BED-defined) centers.
}
\seealso{
Other plotting functions: \code{\link{get_gene_quantiles}},
  \code{\link{plot_percent_meth_with_depth}},
  \code{\link{plot_percent_meth}}
}

