% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methplot.r
\name{plot_percent_meth}
\alias{plot_percent_meth}
\title{Plot binned percent meth table.}
\usage{
plot_percent_meth(binned_perc_meth_table, manual_colors = FALSE)
}
\arguments{
\item{binned_perc_meth_table}{A data.frame or data table.}

\item{manual_colors}{A logical.}
}
\value{
data.table
}
\description{
Plot binned percent meth table.
}
\seealso{
Other plotting functions: \code{\link{get_gene_quantiles}}
}

