% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{plot_percent_meth_with_depth}
\alias{plot_percent_meth_with_depth}
\title{Plot binned percent meth table with a depth (as number of CpGs assayed) facet.}
\usage{
plot_percent_meth_with_depth(binned_perc_meth_table)
}
\arguments{
\item{binned_perc_meth_table}{A data.frame or data table.}
}
\value{
ggplot
}
\description{
Plot binned percent meth table with a depth (as number of CpGs assayed) facet.
}
\seealso{
Other plotting functions: \code{\link{get_gene_quantiles}},
  \code{\link{plot_generic_aggregate_enrichment}},
  \code{\link{plot_percent_meth}}
}

