% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{repeat_mask_meth_table}
\alias{repeat_mask_meth_table}
\title{Mask CpGs in repeats}
\usage{
repeat_mask_meth_table(repeat_mask_ucsc_path, meth_table)
}
\arguments{
\item{repeat_mask_ucsc_path}{A character.}

\item{meth_table}{A data.table.}
}
\value{
data.table
}
\description{
Mask CpGs in repeats
}
\seealso{
Other high level functions: \code{\link{get_tss_perc_meth}}
}

