% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_binned_perc_meth}
\alias{get_binned_perc_meth}
\title{Get percent methylation table by bin.}
\usage{
get_binned_perc_meth(ranges, meth_table)
}
\arguments{
\item{meth_table}{A data.table.}

\item{bin_ranges}{A data.frame.}
}
\value{
data.table
}
\description{
Get percent methylation table by bin.
}
\seealso{
Other workorse functions: \code{\link{genomic_complement}},
  \code{\link{get_bin_ranges_bed}},
  \code{\link{get_bin_ranges_refgene}},
  \code{\link{get_binned_perc_meth_list}},
  \code{\link{get_genome_range_overlaps}},
  \code{\link{get_intersect_single_bases}},
  \code{\link{parallel_genomic_complement}}
}

