% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_genome_range_overlaps}
\alias{get_genome_range_overlaps}
\title{Get range overlaps}
\usage{
get_genome_range_overlaps(bin_ranges, bed_path)
}
\arguments{
\item{bin_ranges}{A data.frame.}

\item{bed_path}{A character.}
}
\value{
data.table
}
\description{
Get range overlaps
}
\seealso{
Other workorse functions: \code{\link{add_chip_enrich_to_binned_perc_meth}},
  \code{\link{genomic_complement}},
  \code{\link{get_aggregate_bed_enrichment_over_ranges}},
  \code{\link{get_bin_ranges_bed}},
  \code{\link{get_bin_ranges_refgene}},
  \code{\link{get_binned_perc_meth_list}},
  \code{\link{get_binned_perc_meth}},
  \code{\link{get_intersect_single_bases}},
  \code{\link{parallel_genomic_complement}}
}

