% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_intersect_single_bases}
\alias{get_intersect_single_bases}
\title{Take BED-formatted single base coords from arbitrary number of tables and return base intersection coords}
\usage{
get_intersect_single_bases(...)
}
\arguments{
\item{...}{A list of meth tables (data.tables).}
}
\value{
A list of data.tables.
}
\description{
Take BED-formatted single base coords from arbitrary number of tables and return base intersection coords
}
\seealso{
Other workorse functions: \code{\link{get_bin_ranges_bed}},
  \code{\link{get_bin_ranges_refgene}},
  \code{\link{get_binned_perc_meth}},
  \code{\link{get_genome_range_overlaps}}
}

