% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methplot.r
\name{get_tss_perc_meth}
\alias{get_tss_perc_meth}
\title{Get percent methylation table around gene model TSSs.}
\usage{
get_tss_perc_meth(refgene_path, meth_table)
}
\arguments{
\item{refgene_path}{A character.}

\item{meth_table}{A data.frame.}
}
\value{
data.table
}
\description{
Get percent methylation table around gene model TSSs.
}

